Extrais `<TagNiveauDroit />`
… pour le réutiliser sur la liste des contributeur à inviter.
Par défaut on invite chaque contributeur en mode `ECRITURE`.
